data/requirements.txt
python-binance==1.0.17
requests==2.28.2
psycopg2==2.9.3